test/api_test/requirements.txt
pytest==7.4.0
requests==2.31.0
PyYAML==6.0.1
pytest-html==4.0.2
allure-pytest==2.13.2